requirements.txt
# Project dependencies
django
django-opensearch
django-bootstrap3
django-bulk-update
pytz
-e git://code.not-your-server.de/pyrpn.git#egg=pyrpn
